requirements.txt
scrapy
pandas
borb
pdfminer.six

unidecode
langdetect